+++ NEWS.md
@@ -1,6 +1,7 @@
 # scorecard 0.3.0.999
 
 * added a parameter in scorecard/scorecard2 function to round score
+* added paramerters show_varval/show_lineval in woebin_plot function to show bar value or line value
 
 # scorecard 0.3.0